+++ algorithms/__init__.py
@@ -16,6 +16,7 @@ structure.Extracted_epochs.RMS = RMS
 structure.Extracted_epochs.Spectrum = Spectrum
 structure.Extracted_epochs.Time_frequency = Time_frequency
 structure.Extracted_epochs.topography = topography  
+structure.Extracted_epochs.frequency_topography = frequency_topography
 structure.Extracted_epochs.significant_channels_count = significant_channels_count
 structure.Extracted_epochs.clustering = clustering
 structure.Extracted_epochs.tanova = tanova