+++ changelog.md
@@ -1,11 +1,20 @@
 # Future works: 
 - correct()
 - Better support for parallel cores
-- Make the "extract()" more flexible
+
+# 0.8.3 (todo)
+- (todo) Make the "extract()" more flexible
+
+# 0.8.2 (2018-03-05)
+- Removed 'time_group' index which are unnecessary in Analyzed_Result
+- Refine X axis in figure
+- Improved the compatibility of save() and load() of Analyzed_Result
+- New algorithm 'frequency_topography()'. Now we can calculate the topography of frequency.
+- New method 'convert()' in module 'Basic'
 
 # 0.8.1 (2018-03-01)
 - MASSIVE amount of changes
-- Altered the package name to "Cafe"
+- Altered the package name to "easyEEG"
 
 # 0.8 (2017-12-25)
 - First to public release. Merry Xmax!